backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
notion-client==2.2.1
python-dotenv==1.0.1
sqlalchemy==2.0.27
pydantic==2.6.1 
pydantic-settings==2.0.0